Relocate spatial package under util package
