+++ util/spatial/doc.go
